Move start modal out of Player component
It should be its own component
